LayoutTests/fast/canvas/webgl/webgl2-buffer-targets-expected.txt

CONSOLE MESSAGE: line 1: TypeError: null is not an object (evaluating 'context.NO_ERROR')
Checks that using WebGL 2 buffer targets don't cause errors

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
